pyproject.toml
[tool.poetry]
name = "z-retina"
version = "0.2.0"
description = "CAFormer-B36 fundus multi-label classification (training + Table S1 eval + patient splits)"
authors = []
readme = "README.md"
packages = [{ include = "z_retina", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
torch = ">=2.4.0"
torchvision = ">=0.19.0"
timm = ">=1.0.0"
albumentations = ">=1.4.0"
opencv-python = ">=4.8.0"
scikit-image = ">=0.22.0"
scikit-learn = ">=1.3.0"
numpy = ">=1.26.0"
pyyaml = ">=6.0"
tqdm = ">=4.66.0"
torch-ema = ">=0.3.0"
iterative-stratification = ">=0.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ipykernel = "^6.29.0"
jupyter = "^1.1.0"

[tool.poetry.scripts]
z-retina-train = "z_retina.apps.train:main"
z-retina-eval = "z_retina.apps.evaluate_cli:main"
z-retina-split = "z_retina.apps.split_cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
